refactor(invitation): consolidate invitation service methods and update imports for improved structure
