+++ web-app/src/app/(app)/organizations/page.tsx
@@ -2,8 +2,7 @@ import { Metadata } from "next";
 import { getTranslations } from "next-intl/server";
 
 import { getSessionOrRedirect } from "@/lib/auth/utils";
-import { MemberService } from "@/lib/services";
-import { getMyValidPendingInvitations } from "@/lib/services/invitation";
+import { InvitationService, MemberService } from "@/lib/services";
 
 import Organizations from "./components/organizations";
 
@@ -21,7 +20,8 @@ export default async function OrganizationsPage() {
   const t = await getTranslations("App.Organizations");
 
   const members = await MemberService.getInstance().getMyMembers();
-  const invitations = await getMyValidPendingInvitations();
+  const invitations =
+    await InvitationService.getInstance().getMyValidPendingInvitations();
 
   return (
     <div className="container flex flex-col gap-8 p-8">
