--- web-app/src/lib/services/invitation/index.ts
@@ -1,2 +1,0 @@
-export * from "./service";
-export * from "./types";
