--- web-app/src/lib/services/invitation/types.ts
@@ -1,5 +1,0 @@
-export enum InvitationErrorCode {
-  EXPIRED = "EXPIRED",
-  NOT_FOUND = "NOT_FOUND",
-  INVITER_NOT_FOUND = "INVITER_NOT_FOUND",
-}
